first commit in branch affaire

+++ db/migrate/20220816162841_create_weektimes.rb
@@ -5,7 +5,7 @@ class CreateWeektimes < ActiveRecord::Migration[7.0]
     create_table :weektimes do |t|
       t.references :user, null: false, foreign_key: true
       t.date :dateweek
-      t.integer :workweek, default: 0
+      t.float :workweek, default: 0
       t.boolean :accord, default: false
 
       t.timestamps
